Refactor - Message Dispatcher

+++ common/message.h
@@ -1,12 +1,32 @@
+// Copyright (c) 2017 david++
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 #ifndef _COMMON_MESSAGE_H
 #define _COMMON_MESSAGE_H
 
 #include "tinyworld.h"
+
 #include <iostream>
 #include <sstream>
 
-////////////////////////////////////////////////////////////
-//
+
 // MessageHeader
 // 
 // +-------------+-------------+
@@ -14,33 +34,30 @@
 // +-------------+-------------+
 //      4bytes       mh.size
 //
-////////////////////////////////////////////////////////////
 
 #pragma pack(1)
 
-struct MessageHeader
-{
-	MessageHeader()
-	{
-		size = 0;
-		zip = 0;
-		encrypt = 0;
-		checksum = 0;
-		reserved = 0;
-	}
+struct MessageHeader {
+    MessageHeader() {
+        size = 0;
+        zip = 0;
+        encrypt = 0;
+        checksum = 0;
+        reserved = 0;
+    }
 
-    // 整个消息的大小
-    size_t msgsize() { return sizeof(MessageHeader) + size; }
+    // whole message size
+    uint32_t msgsize() { return sizeof(MessageHeader) + size; }
 
-    // 消息体的大小
-    size_t bodysize() { return size; }
+    // message bodysize
+    uint32_t bodysize() { return size; }
 
-    std::string dumphex()
-    {
+    // TODO: debug only
+    std::string dumphex() {
         std::ostringstream oss;
-        uint8* bin = (uint8*)this;
-        for (size_t i = 0; i < sizeof(MessageHeader); ++ i)
-            oss << (int)bin[i] << ",";
+        uint8_t *bin = (uint8_t *) this;
+        for (size_t i = 0; i < sizeof(MessageHeader); ++i)
+            oss << (int) bin[i] << ",";
 
         oss << " | size=" << size
             << " zip=" << zip
@@ -50,33 +67,29 @@ struct MessageHeader
         return oss.str();
     }
 
-    uint32 size:24;            // message body size
-	uint32 zip:1;              // message is zipped
-	uint32 encrypt:1;          // message is encrypted
-	uint32 checksum:1;         // undefined
-    uint32 type_is_name:1;     // type is name
-	uint32 reserved:4;         // undefined
+    uint32_t size:24;            // message body size
+    uint32_t zip:1;              // message is zipped
+    uint32_t encrypt:1;          // message is encrypted
+    uint32_t checksum:1;         // undefined
+    uint32_t type_is_name:1;     // type is name
+    uint32_t reserved:4;         // undefined
 
-	union {
-		struct {
-			uint8 type_first;  // first type 
-			uint8 type_second; // second type
-		};
-		uint16 type;           // message type;
-
-        uint16 type_len;       // message type length(when type_is_name=1)
-	};
+    union {
+        struct {
+            uint8_t type_first;  // first type
+            uint8_t type_second; // second type
+        };
+        uint16_t type;           // message type;
+        uint16_t type_len;       // message type length(when type_is_name=1)
+    };
 };
-
 
 #pragma pack()
 
-//
-// 消息类型的工具函数
-//
 #define MSG_TYPE_1(type) ((type) & 0xff)
 #define MSG_TYPE_2(type) ((type) >> 8)
 #define MAKE_MSG_TYPE(type1, type2) (((type2) << 8) + (type1))
+
 //inline uint8 MSG_TYPE_1(const uint16 type)
 //{
 //    MessageType mt;
